change isAdmin boolean->Boolean, personalNumber Integer->Long

+++ persistence/src/main/java/cz/muni/fi/pa165/travelagency/persistence/dao/UserDao.java
@@ -35,7 +35,7 @@ public interface UserDao extends GenericDao<User, Long> {
      * @param personalNumber to find
      * @return list of users with given personal number
      */
-    List<User> findByPersonalNumber(Integer personalNumber);
+    List<User> findByPersonalNumber(Long personalNumber);
 
     /**
      * Method to find all Users with given mail
